requirements.txt
python-socketio
aiohttp
bcrypt
aiosqlite